refactor(backoffice-v2): moved loaders to separate files
