+++ apps/backoffice-v2/src/Router/Router.tsx
@@ -1,31 +1,23 @@
 import React, { FunctionComponent } from 'react';
 import { env } from '../common/env/env';
-import { createBrowserRouter, Outlet, redirect, RouterProvider } from 'react-router-dom';
+import { createBrowserRouter, Outlet, RouterProvider } from 'react-router-dom';
 import { RootError } from '../routes/Root/Root.error';
 import { Root } from '../routes/Root/Root.page';
 import { SignIn } from '../routes/SignIn/SignIn.page';
-import { authQueryKeys } from '../domains/auth/query-keys';
-import { queryClient } from '../lib/react-query/query-client';
-import { filtersQueryKeys } from '../domains/filters/query-keys';
 import { Entity } from '../routes/Entity/Entity.page';
 import { Entities } from '../routes/Entities/Entities.page';
-import { queryKeys } from '../domains/entities/query-keys';
-import { usersQueryKeys } from '../domains/users/query-keys';
-import { searchParamsToObject } from '../common/hooks/useZodSearchParams/utils/search-params-to-object';
 import { RouteError } from '../common/components/atoms/RouteError/RouteError';
 import { CaseManagement } from '../routes/CaseManagement/CaseManagement.page';
+import { rootLoader } from '../routes/Root/Root.loader';
+import { entityLoader } from '../routes/Entity/Entity.loader';
+import { entitiesLoader } from '../routes/Entities/Entities.loader';
+import { localeLoader } from '../routes/Locale/Locale.loader';
 
 const router = createBrowserRouter([
   {
     path: '/',
     element: <Root />,
-    loader({ request }) {
-      const url = new URL(request.url);
-
-      if (url.pathname.startsWith('/en')) return null;
-
-      return redirect(`/en${url.pathname === '/' ? '' : url.pathname}`);
-    },
+    loader: rootLoader,
     errorElement: <RootError />,
     children: [
       ...(env.VITE_AUTH_ENABLED
@@ -39,22 +31,7 @@ const router = createBrowserRouter([
       {
         path: '/:locale',
         element: <Outlet />,
-        async loader() {
-          if (!env.VITE_AUTH_ENABLED) return null;
-
-          const authenticatedUser = authQueryKeys.authenticatedUser();
-          const session = await queryClient.ensureQueryData(
-            authenticatedUser.queryKey,
-            authenticatedUser.queryFn,
-          );
-
-          if (!session?.user) return null;
-
-          const filtersList = filtersQueryKeys.list();
-          await queryClient.ensureQueryData(filtersList.queryKey, filtersList.queryFn);
-
-          return null;
-        },
+        loader: localeLoader,
         errorElement: <RouteError />,
         children: [
           {
@@ -64,42 +41,13 @@ const router = createBrowserRouter([
               {
                 path: '/:locale/case-management/entities',
                 element: <Entities />,
-                async loader({ request }) {
-                  const url = new URL(request.url);
-                  const { entity, filterId } = searchParamsToObject(url.searchParams);
-
-                  if (!entity || !filterId) return null;
-
-                  const entityList = queryKeys[entity].list?.(filterId);
-                  const usersList = usersQueryKeys.list();
-                  await queryClient.ensureQueryData(entityList.queryKey, entityList.queryFn);
-                  await queryClient.ensureQueryData(usersList.queryKey, usersList.queryFn);
-
-                  return null;
-                },
+                loader: entitiesLoader,
                 errorElement: <RouteError />,
                 children: [
                   {
                     path: '/:locale/case-management/entities/:entityId',
                     element: <Entity />,
-                    async loader({ params, request }) {
-                      const url = new URL(request.url);
-                      const { entityId } = params;
-                      const entity = url?.searchParams?.get('entity');
-                      const filterId = url?.searchParams?.get('filterId');
-
-                      if (entity || !filterId) return null;
-
-                      const entityById = queryKeys[entity].byId(entityId, filterId);
-                      // TODO: Add workflowId to params/searchParams
-                      // const workflowById = workflows.byId({ workflowId });
-
-                      await queryClient.ensureQueryData(entityById.queryKey, entityById.queryFn);
-
-                      // await queryClient.ensureQueryData(workflowById.queryKey, workflowById.queryFn);
-
-                      return null;
-                    },
+                    loader: entityLoader,
                     errorElement: <RouteError />,
                   },
                 ],
