+++ apps/backoffice-v2/src/domains/auth/components/UnauthenticatedLayout/UnauthenticatedLayout.layout.tsx
@@ -1,8 +1,7 @@
-import { useUnauthenticatedLayout } from './hooks/useUnauthenticatedLayout/useUnauthenticatedLayout';
 import { FunctionComponentWithChildren } from '../../../../common/types';
 
 export const UnauthenticatedLayout: FunctionComponentWithChildren = ({ children }) => {
-  useUnauthenticatedLayout();
+  // useUnauthenticatedLayout();
 
   return (
     <main className={`h-full`}>
