+++ apps/backoffice-v2/src/domains/workflows/hooks/queries/useWorkflowQuery/useWorkflowQuery.tsx
@@ -2,11 +2,14 @@ import { useQuery } from '@tanstack/react-query';
 import { createWorkflow } from '@ballerine/workflow-browser-sdk';
 import { isString } from '../../../../../common/utils/is-string/is-string';
 import { workflowsQueryKeys } from '../../../query-keys';
+import { useIsAuthenticated } from '../../../../auth/context/AuthProvider/hooks/useIsAuthenticated/useIsAuthenticated';
 
 export const useWorkflowQuery = ({ workflowId }: { workflowId: string }) => {
+  const isAuthenticated = useIsAuthenticated();
+
   return useQuery({
     ...workflowsQueryKeys.byId({ workflowId }),
-    enabled: isString(workflowId) && !!workflowId.length,
+    enabled: isString(workflowId) && !!workflowId.length && isAuthenticated,
     select: ({ workflowDefinition, workflowRuntimeData }) => {
       const { definition, definitionType, ...rest } = workflowDefinition;
       const workflow = {
