+++ apps/backoffice-v2/src/routes/Root/Root.loader.ts
@@ -1,0 +1,9 @@
+import { LoaderFunction, redirect } from 'react-router-dom';
+
+export const rootLoader: LoaderFunction = ({ request }) => {
+  const url = new URL(request.url);
+
+  if (url.pathname.startsWith('/en')) return null;
+
+  return redirect(`/en${url.pathname === '/' ? '' : url.pathname}`);
+};
